requirements.txt
-r requirements-dev.txt
gunicorn
psycopg2==2.9.3
whitenoise-6.2.0